+++ html2text/__init__.py
@@ -32,7 +32,7 @@ except NameError:
     # python3 uses chr
     nochr = str("")
 
-__version__ = (2018, 1, 9)
+__version__ = (2019, 8, 11)
 
 
 # TODO: