+++ ChangeLog.rst
@@ -15,6 +15,7 @@
   * ``html2text.HTML2Text.replaceEntities()``
   * ``html2text.HTML2Text.unescape()``
   * ``html2text.unescape()``
+* Fix #208: handle LEFT-TO-RIGHT MARK after a stressed tag.
 
 2018.1.9
 ========