+++ ChangeLog.rst
@@ -1,5 +1,5 @@
-0000.0.0
-========
+2019.8.11
+=========
 ----
 
 * Add support for wrapping list items.